fix - set autofocus on first input element on each section in event creation form
When user clicks next on the Create Event form, we should also auto focus on the first input on the next form section.
Current use of autofocus attributes won't trigger cursor focus properly. This usually happens on single page app because autofocus attribute only works on page load. But with switching between views or templates in a single page app, the autofocus attribute is not able to trigger input focus.
Hence, used the javascript DOM object to implement this  feature to set the forcus after the section is displyed.