fix - add event button click should show the first section
Earlier whenever the add event button was clicked, it used to show the form section where it was left previously. It also used to show the previous error messages if any. Now fixed that error with resettign the validation messages. Also, whenever the add event button is clicked, it always shows a fresh form with first section and clears the previous error messages